rev1 - Initial commit
To honor one year since the passing of our father, Touraj Ghaffari, we
as his three sons and the most important acheivement in his life,
release to the public his Active Wave active RFID system, the second
most important acheivement in his life. This will be the first of
several commits to upload the entirety of the system.